src/requirements.txt
pandas
pyaml
scikit-learn
scipy